Fix broken lines bug
Previously there was a bug where when the user moved too quickly or in a
strange way, the line would not know how to render, since the 2 points
in the points array weren't adjacent. Now, if the point isn't adjacent,
it just creates a new line, solving this problem once and for all.

Also fixes an issue where going out of the screen and then back in would
cause the program to not know how to render the line. There is no
"staying at the same place but still getting a new point in the points
list" tile. Simply ignoring mouseover events on the last point fixes
this