Update line selection logic
Loops through the lines at the clicked square in the selection for every
click. This makes every line accessible, even if it's an early line that
got covered up by other lines.

Clicking on a line that is already selected or is at the last available
selection deselects it. When this wasn't the case, something just felt
*off* about the interaction